refactor: move line change schemas to shared
